fdgdf

+++ Grand Hotel/ConsoleApp1/Facture.cs
@@ -26,7 +26,7 @@ namespace BOL
     public class LigneFactureBOL
     {
         public int NumLigne { get; set; }
-        public int Quantite { get; set; }
+        public short Quantite { get; set; }
         public decimal MontantHT { get; set; }
         public decimal TauxTVA { get; set; }
         public decimal TauxReduction { get; set; }
